So much refactoring, bro. Very nice.
