+++ src/civilization/census/mod.rs
@@ -3,3 +3,11 @@ pub mod census_systems;
 pub mod census_events;
 pub mod census_components;
 pub mod census_resources;
+
+pub mod prelude {
+    pub use super::census_components::*;
+    pub use super::census_events::*;
+    pub use super::census_plugin::*;
+    pub use super::census_systems::*;
+    pub use super::census_resources::*;
+}
