+++ src/civilization/city_construction/mod.rs
@@ -1,4 +1,11 @@
 pub mod city_construction_plugin;
 pub mod city_construction_components;
-mod city_construction_systems;
+pub mod city_construction_systems;
 pub mod city_construction_events;
+
+pub mod prelude {
+    pub use super::city_construction_components::*;
+    pub use super::city_construction_events::*;
+    pub use super::city_construction_plugin::*;
+    pub use super::city_construction_systems::*;
+}
