+++ src/civilization/city_support/mod.rs
@@ -2,3 +2,10 @@ pub mod city_support_plugin;
 pub mod city_support_events;
 pub mod city_support_systems;
 pub mod city_support_components;
+
+pub mod prelude {
+    pub use super::city_support_components::*;
+    pub use super::city_support_events::*;
+    pub use super::city_support_plugin::*;
+    pub use super::city_support_systems::*;
+}
