+++ src/civilization/conflict/mod.rs
@@ -2,3 +2,10 @@ pub mod conflict_plugin;
 pub mod conflict_components;
 pub mod conflict_systems;
 pub mod conflict_triggers;
+
+pub mod prelude {
+    pub use super::conflict_components::*;
+    pub use super::conflict_plugin::*;
+    pub use super::conflict_systems::*;
+    pub use super::conflict_triggers::*;
+}
