+++ src/civilization/game_moves/mod.rs
@@ -3,3 +3,11 @@ pub mod game_moves_triggers;
 pub mod game_moves_events;
 pub mod game_moves_components;
 pub mod game_moves_systems;
+
+pub mod prelude {
+    pub use super::game_moves_components::*;
+    pub use super::game_moves_events::*;
+    pub use super::game_moves_plugin::*;
+    pub use super::game_moves_systems::*;
+    pub use super::game_moves_triggers::*;
+}
