+++ src/civilization/movement/mod.rs
@@ -2,3 +2,10 @@ pub mod movement_plugin;
 pub mod movement_systems;
 pub mod movement_components;
 pub mod movement_events;
+
+pub mod prelude {
+    pub use super::movement_components::*;
+    pub use super::movement_events::*;
+    pub use super::movement_plugin::*;
+    pub use super::movement_systems::*;
+}
