+++ src/civilization/remove_surplus/mod.rs
@@ -1,2 +1,7 @@
 pub mod remove_surplus_plugin;
 pub mod remove_surplus_systems;
+
+pub mod prelude {
+    pub use super::remove_surplus_plugin::*;
+    pub use super::remove_surplus_systems::*;
+}
