+++ src/stupid_ai/mod.rs
@@ -3,3 +3,11 @@ pub mod stupid_ai_triggers;
 pub mod stupid_ai_events;
 pub mod stupid_ai_components;
 pub mod stupid_ai_systems;
+
+pub mod prelude {
+    pub use super::stupid_ai_components::*;
+    pub use super::stupid_ai_events::*;
+    pub use super::stupid_ai_plugin::*;
+    pub use super::stupid_ai_systems::*;
+    pub use super::stupid_ai_triggers::*;
+}
