+++ src/stupid_ai/stupid_ai_plugin.rs
@@ -1,5 +1,4 @@
-use crate::stupid_ai::stupid_ai_events::{SelectStupidMove, StupidAiEvent};
-use crate::stupid_ai::{stupid_ai_systems, stupid_ai_triggers};
+use crate::stupid_ai::prelude::*;
 use crate::GameState;
 use bevy::app::{Plugin, Update};
 use bevy::prelude::{in_state, App, IntoSystemConfigs};
@@ -12,12 +11,12 @@ impl Plugin for StupidAiPlugin {
             .add_event::<StupidAiEvent>()
             .add_event::<SelectStupidMove>()
             .add_systems(Update, (
-                stupid_ai_systems::setup_stupid_ai.run_if(in_state(GameState::Playing)),
-                stupid_ai_systems::select_stupid_move
+                setup_stupid_ai.run_if(in_state(GameState::Playing)),
+                select_stupid_move
                     .run_if(
                         in_state(GameState::Playing))),
             )
-            .observe(stupid_ai_triggers::on_add_available_moves)
+            .observe(on_add_available_moves)
         ;
     }
 }
