+++ tests/population_expansion_specs.rs
@@ -4,10 +4,11 @@ use crate::common::{create_area, setup_bevy_app, setup_player};
 use adv_civ::civilization::game_moves::game_moves_components::{AvailableMoves, Move};
 use adv_civ::civilization::game_moves::game_moves_events::RecalculatePlayerMoves;
 use adv_civ::civilization::game_moves::game_moves_systems::recalculate_pop_exp_moves_for_player;
-use adv_civ::civilization::general::general_components_file::{PlayerAreas, PlayerStock, Population};
+use adv_civ::civilization::general::general_components::*;
 use adv_civ::civilization::general::general_enums::GameFaction;
 use bevy::app::Update;
 use bevy::prelude::Events;
+use adv_civ::civilization::general::prelude::*;
 
 #[test]
 fn calculate_game_moves_in_population_expansion() {
